requirements.txt
albumentations==0.5.2
gdown==3.13.0
geffnet==1.0.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
matplotlib==3.4.2
numpy==1.20.3
oauthlib==3.1.1
opencv-contrib-python==4.5.2.52
pandas==1.2.4
Pillow==8.2.0
requests==2.25.1
requests-oauthlib==1.3.0
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.5